+++ 3dprog3Project/Src/Graphics/MainRenderPass.cpp
@@ -206,7 +206,7 @@ MainRenderPass::MainRenderPass(ID3D12Device* device, int framesInFlight, DXGI_FO
 	rootSignDesc.Flags = D3D12_ROOT_SIGNATURE_FLAG_NONE;
 	//rtx 2070 mobile does only have shader model 6.5
 	//rootSignDesc.Flags = D3D12_ROOT_SIGNATURE_FLAG_CBV_SRV_UAV_HEAP_DIRECTLY_INDEXED;
-	rootSignDesc.NumParameters = rootParameters.size();
+	rootSignDesc.NumParameters = static_cast<UINT>(rootParameters.size());
 	rootSignDesc.pParameters = rootParameters.data();
 	rootSignDesc.NumStaticSamplers = 1;
 	rootSignDesc.pStaticSamplers = &staticSampler;
@@ -277,7 +277,7 @@ MainRenderPass::~MainRenderPass()
 RenderPassRequirements MainRenderPass::GetRequirements()
 {
 
-	int perThreadSize = rfe::EntityReg::ViewEntities<MeshComp, MaterialComp, TransformComp>().size() / m_numThreads;
+	int perThreadSize = static_cast<int>(rfe::EntityReg::ViewEntities<MeshComp, MaterialComp, TransformComp>().size()) / m_numThreads;
 	perThreadSize += rfe::EntityReg::ViewEntities<MeshComp, MaterialComp, TransformComp>().size() % m_numThreads;
 	RenderPassRequirements req;
 	req.cmdListCount = m_numThreads;
@@ -358,7 +358,7 @@ void MainRenderPass::RunRenderPass(std::vector<ID3D12GraphicsCommandList*> cmdLi
 	else
 	{
 		std::vector<std::future<void>> asyncJobs;
-		int segmentSize = entities.size() / m_numThreads;
+		int segmentSize = static_cast<int>(entities.size()) / m_numThreads;
 		for (int i = 0; i < m_numThreads; i++)
 		{
 			int rest = 0;
@@ -426,7 +426,7 @@ static void Draw(int id, ID3D12Device * device, ID3D12GraphicsCommandList * cmdL
 	uint64_t nextMeshIndexCount = 0;
 	uint64_t nextMeshVertexStart = 0;
 	uint64_t nextMeshVertexCount = 0;
-	int numEntitiesToDraw = entitiesToDraw.size();
+	int numEntitiesToDraw = static_cast<int>(entitiesToDraw.size());
 
 	for (int i = 0; i < numEntitiesToDraw; i++)
 	{
@@ -495,6 +495,6 @@ void MainRenderPass::UpdateDynamicLights(int frameIndex)
 	//this could be done better, we uppdate all lights even if they have not changed
 	if (!lights.empty())
 	{
-		m_dynamicPointLightBuffer[frameIndex]->Update(lights.data(), lights.size());
+		m_dynamicPointLightBuffer[frameIndex]->Update(lights.data(), static_cast<UINT>(lights.size()));
 	}
 }
